refactor(core): extract maintenance/tag-helpers.ts — pure tag-derivation helpers
